requirements.txt
feedparser==6.0.10
slack-sdk==3.26.1
requests==2.31.0
python-dotenv==1.0.0 